requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.1.0
by==0.0.4
certifi==2022.9.24
chainmap==1.0.3
charset-normalizer==2.1.1
click==8.1.3
combomethod==1.0.12
driver==0.1.0
exceptiongroup==1.0.1
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
nulltype==2.3.1
options==1.4.10
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
pytest-rerunfailures==10.3
pytest-xdist==3.1.0
python-dotenv==0.21.0
requests==2.28.1
selenium==4.6.0
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
